refactor(#2055): decouple cache/ brick from nexus.core — zero import violations
Eliminate all 7 nexus.core import violations from the cache/ brick,
aligning it with the NEXUS LEGO Architecture (Tier 2 bricks must have
zero imports from kernel internals).

Phase 1: Move CacheStoreABC + NullCacheStore to bricks/cache/cache_store.py
  - Canonical home for the Fourth Pillar ABC (like Backend in backends/)
  - core/cache_store.py becomes backward-compat re-export
  - Eliminates 4/7 violations (domain, inmemory, dragonfly, factory)

Phase 2: Update bricks/cache/ imports to sibling cache_store.py
  - domain.py, dragonfly.py, inmemory.py, factory.py use sibling imports
  - brick.py: NullCacheStore promoted to module-level, isinstance() fix
  - Eliminates 2/7 violations (NullCacheStore in factory + brick)

Phase 3: Move PostgresPersistentViewStore to storage/
  - Persistence concern with RecordStore affinity, not cache
  - bricks/cache/persistent_view_postgres.py becomes backward-compat shim
  - cache/persistent_view_postgres.py shim updated to point to storage/
  - Eliminates 1/7 violations (PersistentView import)

Phase 4: Export CacheStoreABC + NullCacheStore from __init__.py
  - Package is now self-contained: all pillar types exported

Phase 5: Add import identity tests
  - Verify backward-compat re-exports are identity-preserving (is check)
  - Verify zero runtime nexus.core imports via source inspection
  - Fix subscribe() ABC: AbstractAsyncContextManager (no type:ignore)

All backward-compat re-exports are identity-preserving (same object).
131 unit tests pass, 35 cache E2E tests pass.
