+++ pyproject.toml
@@ -747,7 +747,8 @@ ignore = [
 "src/nexus/backends/*.py" = ["ARG002"]  # Keep unused context params for API compatibility
 "src/nexus/ipc/driver.py" = ["ARG002"]  # Backend ABC override — params required by interface
 "src/nexus/core/ace/trajectory.py" = ["ARG002"]  # Keep unused permission param for API compatibility
-"src/nexus/core/cache_store.py" = ["ARG002"]  # NullCacheStore: Null Object pattern — unused args by design
+"src/nexus/core/cache_store.py" = ["ARG002"]  # Backward-compat shim (canonical home: bricks/cache/cache_store.py)
+"src/nexus/bricks/cache/cache_store.py" = ["ARG002"]  # NullCacheStore: Null Object pattern — unused args by design
 "src/nexus/core/nexus_fs.py" = ["ARG002"]  # Constructor accepts cache params for public API
 "src/nexus/core/distributed_lock.py" = ["ARG002"]  # Interface method signature
 "src/nexus/raft/__init__.py" = ["E402"]  # Conditional imports require try/except after TYPE_CHECKING
