+++ src/nexus/bricks/cache/__init__.py
@@ -40,6 +40,7 @@ from nexus.bricks.cache.base import (
     TigerCacheProtocol,
 )
 from nexus.bricks.cache.brick import CacheBrick
+from nexus.bricks.cache.cache_store import CacheStoreABC, NullCacheStore
 from nexus.bricks.cache.factory import CacheFactory
 from nexus.bricks.cache.inmemory import InMemoryCacheStore
 from nexus.bricks.cache.settings import CacheSettings
@@ -59,6 +60,9 @@ __all__ = [
     "PermissionCacheProtocol",
     "ResourceMapCacheProtocol",
     "TigerCacheProtocol",
+    # Fourth Pillar ABC — canonical home since Issue #2055
+    "CacheStoreABC",
+    "NullCacheStore",
     # CacheStoreABC drivers (for DI into CacheBrick/NexusFS)
     # DragonflyCacheStore — lazy import (use: from nexus.bricks.cache.dragonfly import ...)
     "InMemoryCacheStore",
