+++ src/nexus/bricks/cache/brick.py
@@ -33,6 +33,7 @@ from nexus.bricks.cache.base import (
     ResourceMapCacheProtocol,
     TigerCacheProtocol,
 )
+from nexus.bricks.cache.cache_store import NullCacheStore
 from nexus.bricks.cache.domain import (
     EmbeddingCache,
     PermissionCache,
@@ -78,23 +79,7 @@ class CacheBrick:
             settings: Cache configuration. If None, defaults are used.
             record_store: Optional RecordStoreABC for PostgreSQL cache fallback.
         """
-        # Lazy import to avoid circular deps — NullCacheStore is in core
-        from nexus.bricks.cache.inmemory import InMemoryCacheStore
-
-        # Import NullCacheStore without importing nexus.core at module level
-        _NullCacheStore: type | None = None
-        try:
-            from nexus.core.protocols import NullCacheStore as _NS
-
-            _NullCacheStore = _NS
-        except ImportError:
-            pass
-
-        self._store = (
-            cache_store
-            if cache_store is not None
-            else (_NullCacheStore() if _NullCacheStore else InMemoryCacheStore())
-        )
+        self._store = cache_store if cache_store is not None else NullCacheStore()
         self._settings = settings or CacheSettings(dragonfly_url=None)
         self._record_store = record_store
         self._started = False
@@ -237,7 +222,7 @@ class CacheBrick:
     @property
     def has_cache_store(self) -> bool:
         """Whether a real (non-Null) CacheStoreABC driver is active."""
-        return type(self._store).__name__ != "NullCacheStore"
+        return not isinstance(self._store, NullCacheStore)
 
     # ------------------------------------------------------------------
     # CachingBackendWrapper factory
